Extract Tools into domain-specific objects for readability
