+++ src/main/scala/com/zestia/capsulemcp/server/tools/ToolDescriptionHelper.scala
@@ -1,4 +1,4 @@
-package com.zestia.capsulemcp.server
+package com.zestia.capsulemcp.server.tools
 
 object ToolDescriptionHelper {
   def searchToolDescription(
